Conflicts:
	CMake/hoomd/HOOMDCUDASetup.cmake

+++ hoomd/System.h
@@ -214,7 +214,7 @@ class PYBIND11_EXPORT System
                     if (m_is_variable_period)
                         {
                         pybind11::object pynext = m_update_func(m_n);
-                        int next = (int)pybind11::cast<float>(pynext) + m_created_tstep;
+                        int next = pybind11::cast<int>(pynext) + m_created_tstep;
 
                         if (next < 0)
                             {
@@ -321,7 +321,7 @@ class PYBIND11_EXPORT System
                     if (m_is_variable_period)
                         {
                         pybind11::object pynext = m_update_func(m_n);
-                        int next = (int)pybind11::cast<float>(pynext) + m_created_tstep;
+                        int next = pybind11::cast<int>(pynext) + m_created_tstep;
 
                         if (next < 0)
                             {
